better tests for getTreatmentsRequestPayloadToGateType
